fix: fix case where written digits overlap
In cases where the written digits overlap, the first digit is the one
that "wins" and gets converted. This seems to work against the test they
provide, but it fails against the entire file.

The reason probably has to do with the Replace running against the wrong
location causing confusion instead of a surgical replacement.