feat: write script to download input file for each day
Rather than manually copy/pasting the input file or downloading it
manually, wrote up a small script to download the file and save it to
the year/day location.

This means we need to move the rest of the code and days to follow this
new formatting.

Write up the script with ChatGPT, worked very well!